Organize Storageservices
